+++ src/include/LSP/ClientConfiguration.hpp
@@ -8,8 +8,10 @@ struct ClientDiagnosticsConfiguration
     bool includeDependents = true;
     /// Whether to compute diagnostics for a whole workspace
     bool workspace = false;
+    /// Whether to use expressive DM types in the diagnostics typechecker
+    bool strictDatamodelTypes = true;
 };
-NLOHMANN_DEFINE_TYPE_NON_INTRUSIVE_WITH_DEFAULT(ClientDiagnosticsConfiguration, includeDependents, workspace);
+NLOHMANN_DEFINE_TYPE_NON_INTRUSIVE_WITH_DEFAULT(ClientDiagnosticsConfiguration, includeDependents, workspace, strictDatamodelTypes);
 
 struct ClientSourcemapConfiguration
 {